requirements.txt
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SQLAlchemy==2.4.4
jsonschema==3.2.0
numpy==1.19.4
pandas==1.1.4
plotly==4.14.1
pytest==6.1.2
python-dateutil==2.8.1
psycopg2-binary==2.8.6
requests==2.25.0
simplejson==3.17.2
SQLAlchemy==1.3.20



